View Established
I created a view for the cumulative results, styled it, but still have
to make the controller. I am getting confused when it comes to how to
retrieve some specific data from the database and want to speak to
others about that tomorrow  before I start working on this controller.