Successfully created controller
Changed some things in the model, added a variable to assignment that
correlates to a change in the database where i added a column to the
assignment table called 'dueDate' with a type of 'DATE'. I changed new
class a bit because it is giving trouble for some reason, I will
continue in this branch to fix it.