Completed click functionality + auxiliary changes
I completed the click functionality on the cumulative results page for
students. Additionally, I changed base.html to include the title of the
page (because that was in base_account.html, for some reason). I also
completed the popup function to allow for inputted variables. I added a
GetCriteria function to selection to obtain the criteria which the
selection is in.